requirements.txt
openai==1.3.0
streamlit==1.39.0
requests==2.32.3
pypdf>=4.0.0  # Updated from PyPDF2 to avoid deprecation warning
# tiktoken==0.8.0  # Requires Rust compiler on Python 3.14
numpy>=1.24.0
python-dotenv==1.0.1
plotly>=5.15.0
pandas>=2.2.0
pytest>=7.4.0
pytest-mock>=3.12.0
# chromadb==0.5.23  # Requires onnxruntime which doesn't support Python 3.14 yet